+++ backend/integration_tests/src/update_profile_tests.rs
@@ -23,7 +23,7 @@ fn update_username_succeeds() {
 
     // Check that the user index is updated
     let user_summary = client::user_index::happy_path::users(env, user.principal, canister_ids.user_index, vec![user.user_id])
-        .pop()
+        .current_user
         .unwrap();
     assert_eq!(user_summary.username, username);
 
@@ -54,7 +54,7 @@ fn update_display_name_succeeds() {
 
     // Check that the user index is updated
     let user_summary = client::user_index::happy_path::users(env, user.principal, canister_ids.user_index, vec![user.user_id])
-        .pop()
+        .current_user
         .unwrap();
     assert_eq!(user_summary.display_name, Some(display_name.clone()));